wipm-agent/requirements.txt
fastapi[standard]
latch-sdk-telefonica
mysql-connector-python
python-dotenv